feat: spooled local file and cleanup
