+++ .codespell-whitelist.txt
@@ -1,1 +1,2 @@
 crate
+ser
